Split: take away I3status from __init__.py
